+ fixed a couple of compiler warnings

# Conflicts:
#	.gitignore
#	autostatic.c
#	config.h
#	docs/README.NONPORTABLE
#	pthread.2015.vcxproj
#	pthread.2015.vcxproj.filters
#	pthread.h
#	pthread_delay_np.c
#	pthread_lib.2015.vcxproj
#	pthread_lib.2015.vcxproj.filters
#	pthread_mutex_timedlock.c
#	sched.h
#	semaphore.h

+++ sched_getscheduler.c
@@ -36,7 +36,7 @@
  */
 
 #ifdef HAVE_CONFIG_H
-# include <config.h>
+# include "config.h"
 #endif
 
 #include "pthread.h"
@@ -58,7 +58,7 @@ sched_getscheduler (pid_t pid)
       if (pid != selfPid)
 	{
 	  HANDLE h =
-	    OpenProcess (PROCESS_QUERY_INFORMATION, PTW32_FALSE, (DWORD) pid);
+	    OpenProcess (PROCESS_QUERY_INFORMATION, PTW32_FALSE, pid);
 
 	  if (NULL == h)
 	    {
